+++ examples/selectstream.js
@@ -1,4 +1,4 @@
-/* Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved. */
+/* Copyright (c) 2016, 2020, Oracle and/or its affiliates. All rights reserved. */
 
 /******************************************************************************
  *
@@ -70,6 +70,7 @@ async function run() {
 
       stream.on('end', function() {
         // console.log("stream 'end' event"); // all data has been fetched
+        stream.destroy();                     // clean up resources being used
       });
 
       stream.on('close', function() {